Added isDirectory check on clearing out old script files

+++ server/com.dexels.navajo.compiler.tsl/src/com/dexels/navajo/compiler/tsl/internal/BundleCreatorComponent.java
@@ -213,14 +213,17 @@ public class BundleCreatorComponent implements BundleCreator {
     private void removeOldCompiledScriptFiles(String rpcName) {
         AbstractFileFilter fileFilter = new WildcardFileFilter(FilenameUtils.getBaseName(rpcName) + "*.jar");
         File compiledPath = new File(navajoIOConfig.getCompiledScriptPath(), FilenameUtils.getPath(rpcName));
-        Collection<File> files = FileUtils.listFiles(compiledPath, fileFilter, null);
+        if (compiledPath.isDirectory()) {
+            Collection<File> files = FileUtils.listFiles(compiledPath, fileFilter, null);
 
-        for (File bundleFile : files) {
-            if (bundleFile.isFile() && bundleFile.exists()) {
-                bundleFile.delete();
+            for (File bundleFile : files) {
+                if (bundleFile.isFile() && bundleFile.exists()) {
+                    bundleFile.delete();
+                }
+
             }
-
         }
+      
 
     }
